Add command handling for toss management features
