+++ src/types.ts
@@ -1,0 +1,98 @@
+import { type createReactAgent } from "@langchain/langgraph/prebuilt";
+import { type Client, type Conversation, type DecodedMessage } from "@xmtp/node-sdk";
+
+// Interface for parsed JSON response
+export interface TossJsonResponse {
+  topic?: string;
+  options?: string[];
+  amount?: string;
+  valid?: boolean;
+  reason?: string;
+}
+
+// Interface to track participant options
+export interface Participant {
+  inboxId: string;
+  option: string;
+}
+
+export interface GroupTossName {
+  id: string;
+  creator: string;
+  tossAmount: string;
+  status: TossStatus;
+  participants: string[]; // Maintaining for backward compatibility
+  participantOptions: Participant[]; // New field to track participant options
+  winner?: string;
+  walletAddress: string;
+  createdAt: number;
+  tossResult?: string;
+  paymentSuccess?: boolean;
+  transactionLink?: string;
+  tossTopic?: string;
+  tossOptions?: string[];
+}
+
+export enum TossStatus {
+  CREATED = "CREATED",
+  WAITING_FOR_PLAYER = "WAITING_FOR_PLAYER",
+  READY = "READY",
+  IN_PROGRESS = "IN_PROGRESS",
+  COMPLETED = "COMPLETED",
+  CANCELLED = "CANCELLED",
+}
+
+export interface TransferResponse {
+  model?: {
+    sponsored_send?: {
+      transaction_link?: string;
+    };
+  };
+}
+
+export interface AgentConfig {
+  configurable: {
+    thread_id: string;
+  };
+}
+
+// Interface for parsed toss information
+export interface ParsedToss {
+  topic: string;
+  options: string[];
+  amount: string;
+}
+
+// Define stream chunk types
+export interface AgentChunk {
+  agent: {
+    messages: Array<{
+      content: string;
+    }>;
+  };
+}
+
+export interface ToolsChunk {
+  tools: {
+    messages: Array<{
+      content: string;
+    }>;
+  };
+}
+
+export type StreamChunk = AgentChunk | ToolsChunk;
+
+export type MessageHandler = (
+  message: DecodedMessage,
+  conversation: Conversation,
+  command: string
+) => Promise<void>;
+
+// Interface for transfer response
+export interface Transfer {
+  model?: {
+    sponsored_send?: {
+      transaction_link?: string;
+    };
+  };
+}
